Move nvimproc files to src/, delete directory
