+++ src/Stdout.ts
